+++ src/surrealist/ql/database.py
@@ -5,13 +5,12 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 from surrealist.connections.connection import Connection
 from surrealist.enums import Algorithm, AutoOrNone
 from surrealist.errors import SurrealConnectionError
-from surrealist.ql.statements import (DefineAccessJwt, DefineAccessBearer, DefineAccessRecord)
+from surrealist.ql.statements.define_field import DefineField
+from surrealist.ql.statements.define_access import DefineAccessJwt, DefineAccessBearer, DefineAccessRecord
+from surrealist.ql.statements.define import (DefineEvent, DefineIndex, DefineParam, DefineScope, DefineTable,
+                                             DefineToken)
 from surrealist.ql.statements import Live, Remove, Select
 from surrealist.ql.statements.alter import Alter
-from surrealist.ql.statements.define import (DefineEvent, DefineField,
-                                             DefineIndex, DefineParam,
-                                             DefineScope, DefineTable,
-                                             DefineToken)
 from surrealist.ql.statements.define_analyzer import DefineAnalyzer
 from surrealist.ql.statements.define_config import DefineConfig
 from surrealist.ql.statements.define_user import DefineUser